refactor: move enemy forces to map manager
